Fix high-complexity unused code
