+++ app/src/main/AndroidManifest.xml
@@ -12,19 +12,18 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.EmotionDiary"
         tools:targetApi="31">
-        <activity android:name=".ui.Remind.RemindActivity" />
         <activity
             android:name=".MainActivity"
-            android:exported="true"
-            android:label="@string/app_name">
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-<!--        intent 이동하려면 필요하다해서 추가했습니다-->
-        <activity android:name=".write.DiaryWriteActivity"/>
+
+        <activity android:name=".ui.Remind.RemindActivity" />
+        <activity android:name=".ui.writePage.DiaryWriteActivity"/>
     </application>
 
 </manifest>
