+++ app/src/main/java/kr/co/gachon/emotion_diary/data/AppDatabase.java
@@ -7,11 +7,10 @@ import androidx.room.Room;
 import androidx.room.RoomDatabase;
 import androidx.room.TypeConverters; // Date 타입 변환을 위해 추가
 
-@Database(entities = {Diary.class}, version = 2)
-@TypeConverters({Converters.class}) // Date type changer
+@Database(entities = {Diary.class}, version = 3, exportSchema = false)
+@TypeConverters({Converters.class})
 public abstract class AppDatabase extends RoomDatabase {
-    public abstract DiaryDao diaryDao(); // DiaryDao 추가
-
+    public abstract DiaryDao diaryDao();
     private static volatile AppDatabase INSTANCE;
 
     public static AppDatabase getDatabase(final Context context) {
