+++ app/src/main/res/layout/activity_diary_write.xml
@@ -1,5 +1,49 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/fragment_container"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent" />
+    android:layout_height="match_parent"
+    android:layout_marginTop="100dp"
+    android:orientation="vertical">
+
+    <TextView
+        android:id="@+id/dateTextView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="50dp"
+        android:text="날짜" />
+
+    <EditText
+        android:id="@+id/titleTextView"
+        android:layout_width="match_parent"
+        android:layout_height="50dp"
+        android:hint="제목"
+        android:textSize="20sp" />
+
+    <EditText
+        android:id="@+id/contentTextView"
+        android:layout_width="match_parent"
+        android:layout_height="200dp"
+        android:hint="작성하시오" />
+
+    <TextView
+        android:id="@+id/emotionTextView"
+        android:layout_width="match_parent"
+        android:layout_height="0dp"
+        android:text="감정" />
+
+
+    <Button
+        android:id="@+id/insertDummyData"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="emtion select" />
+
+    <Button
+        android:id="@+id/back_button"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="50dp"
+        android:background="@color/black"
+        android:text="back" />
+
+</LinearLayout>
