--- app/src/main/res/layout/fragment_diary_write.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_marginTop="100dp">
-
-    <FrameLayout
-        android:id="@+id/fragment_container"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"/>
-
-    <TextView
-        android:id="@+id/dateTextView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="50dp"
-        android:text="날짜"/>
-
-    <EditText
-        android:id="@+id/titleTextView"
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:textSize="20sp"
-        android:hint="제목" />
-
-    <EditText
-        android:id="@+id/contentTextView"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:hint="작성하시오" />
-
-    <TextView
-        android:id="@+id/emotionTextView"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:text="감정"/>
-
-
-
-    <Button
-        android:id="@+id/insertDummyData"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="emtion select"
-        />
-
-    <Button
-        android:id="@+id/back_button"
-        android:layout_marginTop="50dp"
-        android:background="@color/black"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="back"
-        />
-
-</LinearLayout>
